feat: confirm password before to change it

+++ src/screens/SignIn.tsx
@@ -17,10 +17,18 @@ export interface FormSignIn extends FormEmail {
   password: string
 }
 
+export interface SignInParams {
+  user: FormEmail
+  isConfirmPassword?: boolean
+}
+
 export default function SignIn() {
   const { navigate } = useNavigation()
   const route = useRoute()
-  const { email } = route.params as FormEmail
+  const {
+    user: { email },
+    isConfirmPassword
+  } = route.params as SignInParams
   const [show, setShow] = React.useState(false)
   const { setUser } = useAuth()
 
@@ -32,19 +40,26 @@ export default function SignIn() {
       setSubmitting(true)
       const response = await api.post('/sign-in', values)
       if (response.data.status && response.data.token) {
-        await AsyncStorage.setItem('accessToken', response.data.token)
-        await AsyncStorage.setItem('user', JSON.stringify(response.data.user))
-        setUser({
-          ...response.data.user,
-          noRedirect: true
-        })
-        navigate('Groups')
+        if (!isConfirmPassword) {
+          await AsyncStorage.setItem('accessToken', response.data.token)
+          await AsyncStorage.setItem('user', JSON.stringify(response.data.user))
+          setUser({
+            ...response.data.user,
+            noRedirect: true
+          })
+          navigate('Groups')
+        } else
+          navigate('Password', {
+            user: response.data.user
+          })
       } else if (response.data.hasOwnProperty('error')) {
         switch (response.data.error) {
           case 'INVALID_PASSWORD':
             Alert.alert(
               'Ops!',
-              'Senha inválida! Tente novamente ou redefina sua senha.'
+              isConfirmPassword
+                ? 'Senha inválida!'
+                : 'Senha inválida! Tente novamente ou redefina sua senha.'
             )
             break
           case 'USER_DOES_NOT_EXIST':
@@ -93,7 +108,9 @@ export default function SignIn() {
               <BackButton />
             </Box>
             <Text my={4} fontSize={28} color="white">
-              Agora sua senha
+              {isConfirmPassword
+                ? 'Confirme sua senha atual'
+                : 'Agora sua senha'}
             </Text>
             <VStack space={8}>
               <TextField
@@ -109,19 +126,21 @@ export default function SignIn() {
                   </Pressable>
                 }
               />
-              <Center>
-                <Pressable
-                  onPress={() => navigate('PasswordRecovery', { email })}
-                >
-                  <Text color="white" underline fontSize="md">
-                    Esqueci minha senha
-                  </Text>
-                </Pressable>
-              </Center>
+              {!isConfirmPassword && (
+                <Center>
+                  <Pressable
+                    onPress={() => navigate('PasswordRecovery', { email })}
+                  >
+                    <Text color="white" underline fontSize="md">
+                      Esqueci minha senha
+                    </Text>
+                  </Pressable>
+                </Center>
+              )}
             </VStack>
           </VStack>
           <SubmitButton
-            title="Entrar"
+            title={isConfirmPassword ? 'Continuar' : 'Entrar'}
             isSubmitting={isSubmitting}
             handleSubmit={handleSubmit}
           />
